[uniffi] Move `GroupStateStorage` wrapper to client
This type is specific to our client type so it can be nicely grouped
with it in the client module.
