+++ mls-rs-uniffi/src/config/group_state.rs
@@ -1,4 +1,4 @@
-use std::{fmt::Debug, sync::Arc};
+use std::fmt::Debug;
 
 use crate::Error;
 
@@ -53,44 +53,4 @@ pub trait GroupStateStorage: Send + Sync + Debug {
     ) -> Result<(), Error>;
 
     async fn max_epoch_id(&self, group_id: Vec<u8>) -> Result<Option<u64>, Error>;
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct GroupStateStorageWrapper(Arc<dyn GroupStateStorage>);
-
-impl From<Arc<dyn GroupStateStorage>> for GroupStateStorageWrapper {
-    fn from(value: Arc<dyn GroupStateStorage>) -> Self {
-        Self(value)
-    }
-}
-
-#[cfg_attr(not(mls_build_async), maybe_async::must_be_sync)]
-#[cfg_attr(mls_build_async, maybe_async::must_be_async)]
-impl mls_rs_core::group::GroupStateStorage for GroupStateStorageWrapper {
-    type Error = Error;
-
-    async fn state(&self, group_id: &[u8]) -> Result<Option<Vec<u8>>, Self::Error> {
-        self.0.state(group_id.to_vec())
-    }
-
-    async fn epoch(&self, group_id: &[u8], epoch_id: u64) -> Result<Option<Vec<u8>>, Self::Error> {
-        self.0.epoch(group_id.to_vec(), epoch_id)
-    }
-
-    async fn write(
-        &mut self,
-        state: mls_rs_core::group::GroupState,
-        inserts: Vec<mls_rs_core::group::EpochRecord>,
-        updates: Vec<mls_rs_core::group::EpochRecord>,
-    ) -> Result<(), Self::Error> {
-        self.0.write(
-            state.into(),
-            inserts.into_iter().map(Into::into).collect(),
-            updates.into_iter().map(Into::into).collect(),
-        )
-    }
-
-    async fn max_epoch_id(&self, group_id: &[u8]) -> Result<Option<u64>, Self::Error> {
-        self.0.max_epoch_id(group_id.to_vec())
-    }
 }
